pyproject.toml
[tool.poetry]
name = "museums"
version = "0.1.0"
description = ""
authors = ["Marius <mariuskrasuckas@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.14"
python-dotenv = "^1.1.1"
pinecone = "^7.3.0"
streamlit = "^1.48.1"
streamlit-clickable-images = "^0.0.3"
langchain-openai = "^0.3.32"
langchain = "^0.3.27"
langchain-pinecone = "^0.2.11"
openai = "^1.102.0"
langchain-community = "^0.3.29"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
